Add build script, bug fixes, TABLES support, and updated docs.
- build.sh: auto-detects MinGW cross-compiler, checks zlib.net for newer
  zlib, downloads/extracts if needed, then cross-compiles oddie.exe
- .gitignore: excludes zlib source files (fetched at build time) and oddie.exe
- md5.h: fix uint32 typedef (unsigned long → uint32_t via stdint.h)
- oddie.c: add DELETE_FILE macro (DeleteFile/unlink portability); widen
  col_name buffer to 128 bytes; add TABLES=1 support via SQLTables; add
  input length guards with ERROR responses; rename length → result_char_len;
  fix buffer-overflow checks (> → >=); improve MariaDB workaround comment
- README.md: document ID/MD5/ZIP/TABLES fields; replace manual gcc command
  with ./build.sh
- docs/: design spec and implementation plan for build script

+++ md5.h
@@ -1,8 +1,8 @@
 #ifndef MD5_H
 #define MD5_H
 
-// xxx add ifdef for 64 bits
-typedef unsigned long uint32;
+#include <stdint.h>
+typedef uint32_t uint32;
 
 typedef struct
 {
